requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
greenlet==1.1.3.post0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.14.0
pluggy==0.13.1
psycopg2==2.9.4
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pytest==4.5.0
six==1.16.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.9.0
